Added/implemented layers
Added/implemented layers to sample scene, added a sprite atlas in the palettes folder and set it up to use the backyard tilesheet. Renamed the already exsisting layers in the samplescene to reflect their sort order. Not sure how you like to name your scenes so you may want to undo the name change.